Decoupled EntitySearcher from Searcher-Implementation by implementing (ES/Lucene)-specific SearchQueryGenerators.
