+++ stargraph-core/src/main/java/net/stargraph/core/IndicesFactory.java
@@ -28,6 +28,7 @@ package net.stargraph.core;
 
 import net.stargraph.core.index.BaseIndexer;
 import net.stargraph.core.search.BaseSearcher;
+import net.stargraph.core.search.SearchQueryGenerator;
 import net.stargraph.model.KBId;
 
 public interface IndicesFactory {
@@ -35,4 +36,6 @@ public interface IndicesFactory {
     BaseIndexer createIndexer(KBId kbId, Stargraph stargraph);
 
     BaseSearcher createSearcher(KBId kbId, Stargraph stargraph);
+
+    SearchQueryGenerator createSearchQueryGenerator(KBId kbId, Stargraph stargraph);
 }
