+++ stargraph-core/src/main/java/net/stargraph/core/search/EntitySearcher.java
@@ -26,25 +26,135 @@ package net.stargraph.core.search;
  * ==========================License-End===============================
  */
 
+import net.stargraph.core.KBCore;
+import net.stargraph.core.Namespace;
+import net.stargraph.model.BuiltInModel;
+import net.stargraph.model.Fact;
 import net.stargraph.model.InstanceEntity;
 import net.stargraph.model.LabeledEntity;
-import net.stargraph.rank.ModifiableRankParams;
-import net.stargraph.rank.ModifiableSearchParams;
-import net.stargraph.rank.Scores;
+import net.stargraph.rank.*;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.slf4j.Marker;
+import org.slf4j.MarkerFactory;
 
+import java.util.Collections;
 import java.util.List;
+import java.util.Objects;
+import java.util.stream.Collectors;
 
-public interface EntitySearcher {
+public class EntitySearcher {
+    private Logger logger = LoggerFactory.getLogger(getClass());
+    private Marker marker = MarkerFactory.getMarker("elastic");
 
-    LabeledEntity getEntity(String dbId, String id);
+    private KBCore core;
 
-    List<LabeledEntity> getEntities(String dbId, List<String> ids);
 
-    Scores classSearch(ModifiableSearchParams searchParams, ModifiableRankParams rankParams);
+    public EntitySearcher(KBCore core) {
+        this.core = Objects.requireNonNull(core);
+    }
 
-    Scores instanceSearch(ModifiableSearchParams searchParams, ModifiableRankParams rankParams);
+    public LabeledEntity getEntity(String dbId, String id) {
+        List<LabeledEntity> res = getEntities(dbId, Collections.singletonList(id));
+        if (res != null && !res.isEmpty()) {
+            return res.get(0);
+        }
+        return null;
+    }
 
-    Scores propertySearch(ModifiableSearchParams searchParams, ModifiableRankParams rankParams);
+    public List<LabeledEntity> getEntities(String dbId, List<String> ids) {
+        logger.info(marker, "Fetching ids={}", ids);
+        Namespace ns = core.getNamespace();
+        List idList = ids.stream().map(ns::shrinkURI).collect(Collectors.toList());
+        ModifiableSearchParams searchParams = ModifiableSearchParams.create(dbId).model(BuiltInModel.ENTITY);
 
-    Scores pivotedSearch(InstanceEntity pivot, ModifiableSearchParams searchParams, ModifiableRankParams rankParams);
+        SearchQueryGenerator searchQueryGenerator = core.getSearchQueryGenerator(searchParams.getKbId().getModel());
+        SearchQueryHolder holder = searchQueryGenerator.entitiesWithIds(idList, searchParams);
+        Searcher searcher = core.getSearcher(searchParams.getKbId().getModel());
+
+        // Fetch initial candidates from the search engine
+        Scores scores = searcher.search(holder);
+
+        return scores.stream().map(s -> (LabeledEntity)s.getEntry()).collect(Collectors.toList());
+    }
+
+    public Scores classSearch(ModifiableSearchParams searchParams, ModifiableRankParams rankParams) {
+
+        searchParams.model(BuiltInModel.FACT);
+
+        if (rankParams instanceof ModifiableIndraParams) {
+            core.configureDistributionalParams((ModifiableIndraParams) rankParams);
+        }
+
+        SearchQueryGenerator searchQueryGenerator = core.getSearchQueryGenerator(searchParams.getKbId().getModel());
+        SearchQueryHolder holder = searchQueryGenerator.findClassFacts(searchParams);
+        Searcher searcher = core.getSearcher(searchParams.getKbId().getModel());
+
+        // Fetch initial candidates from the search engine
+        Scores scores = searcher.search(holder);
+
+        List<Score> classes2Score = scores.stream()
+                .map(s -> new Score(((Fact)s.getEntry()).getObject(), s.getValue())).collect(Collectors.toList());
+        // Re-Rank
+        return Rankers.apply(new Scores(classes2Score), rankParams, searchParams.getSearchTerm());
+    }
+
+    public Scores instanceSearch(ModifiableSearchParams searchParams, ModifiableRankParams rankParams) {
+
+        searchParams.model(BuiltInModel.ENTITY);
+
+        SearchQueryGenerator searchQueryGenerator = core.getSearchQueryGenerator(searchParams.getKbId().getModel());
+        SearchQueryHolder holder = searchQueryGenerator.findEntityInstances(searchParams);
+        Searcher searcher = core.getSearcher(searchParams.getKbId().getModel());
+
+        // Fetch initial candidates from the search engine
+        Scores scores = searcher.search(holder);
+        // Re-Rank
+        return Rankers.apply(scores, rankParams, searchParams.getSearchTerm());
+    }
+
+    public Scores propertySearch(ModifiableSearchParams searchParams, ModifiableRankParams rankParams) {
+
+        searchParams.model(BuiltInModel.PROPERTY);
+
+        if (rankParams instanceof ModifiableIndraParams) {
+            core.configureDistributionalParams((ModifiableIndraParams) rankParams);
+        }
+
+        SearchQueryGenerator searchQueryGenerator = core.getSearchQueryGenerator(searchParams.getKbId().getModel());
+        SearchQueryHolder holder = searchQueryGenerator.findPropertyInstances(searchParams);
+        Searcher searcher = core.getSearcher(searchParams.getKbId().getModel());
+
+        // Fetch initial candidates from the search engine
+        Scores scores = searcher.search(holder);
+        // Re-Rank
+        return Rankers.apply(scores, rankParams, searchParams.getSearchTerm());
+    }
+
+    public Scores pivotedSearch(InstanceEntity pivot,
+                                ModifiableSearchParams searchParams, ModifiableRankParams rankParams) {
+
+        searchParams.model(BuiltInModel.FACT);
+
+        if (rankParams instanceof ModifiableIndraParams) {
+            core.configureDistributionalParams((ModifiableIndraParams) rankParams);
+        }
+
+        SearchQueryGenerator searchQueryGenerator = core.getSearchQueryGenerator(searchParams.getKbId().getModel());
+        SearchQueryHolder holder = searchQueryGenerator.findPivotFacts(pivot, searchParams);
+        Searcher searcher = core.getSearcher(searchParams.getKbId().getModel());
+
+        // Fetch initial candidates from the search engine
+        Scores scores = searcher.search(holder);
+
+        // We have to remap the facts to properties, the real target of the ranker call.
+        // Thus we're discarding the score values from the underlying search engine. Shall we?
+        Scores propScores = new Scores(scores.stream()
+                .map(s -> ((Fact) s.getEntry()).getPredicate())
+                .distinct()
+                .map(p -> new Score(p, 0))
+                .collect(Collectors.toList()));
+
+        return Rankers.apply(propScores, rankParams, searchParams.getSearchTerm());
+    }
 }
